Move sample-list related functionality to lims component
