+++ mxcubeweb/core/components/lims.py
@@ -8,6 +8,10 @@ from mxcubecore.model import queue_model_objects as qmo
 from mxcubecore.model.lims_session import LimsSessionManager
 
 from mxcubeweb.core.components.component_base import ComponentBase
+from mxcubeweb.core.components.queue import (
+    COLLECTED,
+    UNCOLLECTED,
+)
 
 VALID_SAMPLE_NAME_REGEXP = re.compile("^[a-zA-Z0-9:+_-]+$")
 
@@ -32,7 +36,12 @@ class Lims(ComponentBase):
     def sample_list_set_order(self, sample_order):
         self.app.SAMPLE_LIST["sampleOrder"] = sample_order
 
-    def sample_list_get(self, loc=None, current_queue=None):
+    def sample_list_get(
+        self, loc=None, current_queue=None, retrieve_samples_from_sc=False
+    ):
+        if retrieve_samples_from_sc:
+            self.get_sample_list_from_sc()
+
         self.synch_sample_list_with_queue(current_queue)
         res = self.app.SAMPLE_LIST
 
@@ -41,21 +50,112 @@ class Lims(ComponentBase):
 
         return res
 
+    def sc_contents_add(self, sample):
+        code, location = sample.get("code", None), sample.get("sampleID")
+
+        if code:
+            self.app.SC_CONTENTS.get("FROM_CODE")[code] = sample
+        if location:
+            self.app.SC_CONTENTS.get("FROM_LOCATION")[location] = sample
+
+    def sc_contents_from_code_get(self, code):
+        return self.app.SC_CONTENTS["FROM_CODE"].get(code, {})
+
+    def sc_contents_from_location_get(self, loc):
+        return self.app.SC_CONTENTS["FROM_LOCATION"].get(loc, {})
+
+    def set_current_sample(self, sample_id):
+        self.app.CURRENTLY_MOUNTED_SAMPLE = sample_id
+        logging.getLogger("MX3.HWR").info(
+            "[SC] Setting currently mounted sample to %s", sample_id
+        )
+
+        sample_id = sample_id if sample_id else ""
+
+        self.app.server.emit(
+            "set_current_sample", {"sampleID": sample_id}, namespace="/hwr"
+        )
+
+    def get_current_sample(self):
+        return self.app.SAMPLE_LIST["sampleList"].get(
+            self.app.CURRENTLY_MOUNTED_SAMPLE, {}
+        )
+
+    def get_sample_list_from_sc(self):
+        samples_list = (
+            HWR.beamline.sample_changer.get_sample_list()
+            if HWR.beamline.sample_changer
+            else []
+        )
+        samples = {}
+        sample_list_by_coords = {}
+        order = []
+        current_sample = {}
+
+        loaded_sample = (
+            HWR.beamline.sample_changer.get_loaded_sample()
+            if HWR.beamline.sample_changer
+            else None
+        )
+
+        for s in samples_list:
+            if not s.is_present():
+                continue
+            state = COLLECTED if s.has_been_loaded() else UNCOLLECTED
+            sample_dm = s.get_id() or ""
+            coords = s.get_coords()
+            sample_data = {
+                "sampleID": s.get_address(),
+                "location": s.get_address(),
+                "sampleName": s.get_name() or "Sample-%s" % s.get_address(),
+                "crystalUUID": s.get_id() or s.get_address(),
+                "proteinAcronym": (
+                    s.proteinAcronym if hasattr(s, "proteinAcronym") else ""
+                ),
+                "code": sample_dm,
+                "loadable": True,
+                "state": state,
+                "tasks": [],
+                "type": "Sample",
+                "cell_no": s.get_cell_no() if hasattr(s, "get_cell_no") else 1,
+                "puck_no": s.get_basket_no() if hasattr(s, "get_basket_no") else 1,
+            }
+            order.append(coords)
+            sample_list_by_coords[coords] = sample_data["sampleID"]
+
+            sample_data["defaultPrefix"] = self.app.lims.get_default_prefix(sample_data)
+            sample_data["defaultSubDir"] = self.app.lims.get_default_subdir(sample_data)
+
+            samples[s.get_address()] = sample_data
+            self.sc_contents_add(sample_data)
+
+            if loaded_sample and sample_data["location"] == loaded_sample.get_address():
+                current_sample = sample_data
+                self.app.queue.queue_add_item([current_sample])
+
+        # sort by location, using coords tuple
+        order.sort()
+        sample_list = {
+            "sampleList": samples,
+            "sampleOrder": [sample_list_by_coords[coords] for coords in order],
+        }
+
+        self.app.lims.sample_list_set(sample_list)
+
+        if current_sample:
+            self.set_current_sample(current_sample["sampleID"])
+
     def sample_list_sync_sample(self, lims_sample):
         lims_code = lims_sample.get("code", None)
         lims_location = lims_sample.get("lims_location")
         sample_to_update = None
 
         # LIMS sample has code, check if the code was read by SC
-        if lims_code and self.app.sample_changer.sc_contents_from_code_get(lims_code):
-            sample_to_update = self.app.sample_changer.sc_contents_from_code_get(
-                lims_code
-            )
+        if lims_code and self.sc_contents_from_code_get(lims_code):
+            sample_to_update = self.sc_contents_from_code_get(lims_code)
         elif lims_location:
             # Asume that the samples have been put in the right place of the SC
-            sample_to_update = self.app.sample_changer.sc_contents_from_location_get(
-                lims_location
-            )
+            sample_to_update = self.sc_contents_from_location_get(lims_location)
 
         if sample_to_update:
             loc = sample_to_update["sampleID"]
@@ -219,7 +319,7 @@ class Lims(ComponentBase):
             self.app.lims.init_sample_list()
 
             # Get sample list and send update to client
-            self.app.sample_changer.get_sample_list()
+            self.get_sample_list_from_sc()
             self.app.server.emit("update_queue", {}, namespace="/hwr")
 
             HWR.beamline.session.proposal_code = session.code
@@ -270,7 +370,7 @@ class Lims(ComponentBase):
 
     def synch_with_lims(self, lims_name):
         self.app.queue.queue_clear()
-        self.app.sample_changer.get_sample_list()
+        self.get_sample_list_from_sc()
 
         samples_info_list = HWR.beamline.lims.get_samples(lims_name)
         for sample_info in samples_info_list:
